Separate GameService and NewsComponent
