+++ frontend/src/app/game/game/game.component.html
@@ -1,5 +1,7 @@
 <cvd-col [spacing]="32">
 
+  <cvd-news></cvd-news>
+
   <cvd-charts></cvd-charts>
 
   <cvd-row wrap>
