+++ frontend/src/app/game/simulation-control/simulation-control.component.scss
@@ -2,4 +2,5 @@
   min-width: 10rem;
   min-height: 10rem;
   background: bisque;
+  padding: 1rem;
 }
